+++ test/stdlib_basics/test_base.py
@@ -26,5 +26,6 @@ def test_component():
     c = _ClosuredComponent()
     assert len(c.parts()) == 0
 
+
 if __name__ == "__main__":
     pytest.main([__file__])